pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "xcube_gedidb"
dynamic = ["version"]
authors = [
    {name = "Yogesh Kumar Baljeet Singh", email = "yogesh.baljeetsingh@brockmann-consult.de"}
]
description = """\
  xcube datastore for accessing GediDB datastore using the `gedidb` API\
  """
keywords = [
    "analysis ready data", "data science", "datacube", "xarray", "zarr", "xcube", "gedi", "gedidb"
]
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "xarray",
    "xcube-core",
    "gedidb"
]

[tool.setuptools.dynamic]
version = {attr = "xcube_gedidb.__version__"}

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-recording",
    "flake8"
]

[project.urls]
Repository = 'https://github.com/xcube-dev/xcube-gedidb'
Issues = 'https://github.com/xcube-dev/xcube-gedidb/issues'
Changelog = 'https://github.com/xcube-dev/xcube-gedidb/blob/main/CHANGES.md'